pfm/requirements.txt
asgiref==3.8.1
Django==4.2.6
sqlparse==0.5.3
tzdata==2025.1
djangorestframework==3.14.0